+++ NEWS.md
@@ -66,12 +66,33 @@ Added function 'common' which identifies common substrings inside a pair of chai
 
 * Redesigned function "fcommon" which is now fully sequential.
 
-* Expanded function "cover" capability to show a list of valid substrings out of the string covering using
+* Expanded function "cover" capability to show a list of valid substrings out of string covering using
   combinatorial method (recommended for short strings).
 
 * Removed packages 'future.apply' from Depends and 'future' from Suggests in DESCRIPTION.
 
 * Function "common" is now defunct.
 
+# akin 0.3.4
 
+* Minor code optimization for functions "cover" and "fcommon". "cover" now flags when only empty or
+  1-character substrings are identified.
+  
+* Added new example to "fcommon" which proves the completeness of common substring identification.
 
+* Added new example for function "cover" which shows the difference between string covering and valid
+  substrings set.
+
+* Added byte compilation option to DESCRIPTION.
+
+# akin 0.3.5
+
+* Resubmission
+ 
+* Edited "ByteCompile" in the DESCRIPTION file.
+
+# akin 0.3.6
+
+* Resubmission
+
+* Moved package 'RcppAlgos' from Depends to Imports